# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SynthChat Hermes Backend API
  version: 0.1.0
  description: |
    Frontend-facing contract for the complete Rust backend. Upstream Hermes
    HTTP, dashboard, WebSocket, CLI, filesystem and SQLite interfaces are
    compatibility references only and are not production runtime dependencies.
servers:
  - url: http://127.0.0.1:8642
    description: Local desktop backend
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/profiles/{profileId}/session-imports/hermes-v21:
    parameters:
      - $ref: "#/components/parameters/ProfileId"
    get:
      tags: [Sessions]
      operationId: previewHermesV21SessionImport
      summary: Preview the profile's fixed-version Hermes v21 session import
      responses:
        "200":
          description: Source is absent, or a consistent read-only snapshot is ready for explicit import.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HermesImportPreview" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "503": { $ref: "#/components/responses/SessionStorageUnavailable" }
    post:
      tags: [Sessions]
      operationId: importHermesV21Sessions
      summary: Atomically import the previewed Hermes v21 snapshot
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HermesImportRequest" }
      responses:
        "200":
          description: The complete snapshot was imported, was already present, or this idempotent result was replayed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HermesImportResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Idempotency/source conflict, or a bounded atomic-import conflict report.
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/HermesImportConflictProblem"
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "503": { $ref: "#/components/responses/SessionStorageUnavailable" }
  /api/v1/files:
    post:
      tags: [Files]
      operationId: uploadFile
      summary: Upload a file
      description: Stores one opaque file snapshot under HERMES_HOME without accepting a filesystem path. The semantic request fingerprint covers the validated filename, normalized MIME type and complete bytes.
      x-max-file-bytes: 8388608
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: File accepted and represented by an opaque ID.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FileRef" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "507": { $ref: "#/components/responses/InsufficientStorage" }
  /api/v1/files/{fileId}/content:
    parameters:
      - $ref: "#/components/parameters/FileId"
    get:
      tags: [Files]
      operationId: getFileContent
      summary: Read file content
      responses:
        "200":
          description: Immutable file bytes. Content-Type is the normalized allowlisted mimeType returned by upload; responses are no-store and nosniff.
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/v1/files/{fileId}:
    parameters:
      - $ref: "#/components/parameters/FileId"
    delete:
      tags: [Files]
      operationId: deleteFile
      summary: Delete a temporary file
      responses:
        "204": { description: File deleted or already absent. A retained idempotency record is not deleted and will never recreate an explicitly deleted file. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-desktop-session-token
  headers:
    ETag:
      description: Opaque resource revision, including quotes as required by HTTP.
      schema: { type: string }
  parameters:
    ProfileId: { name: profileId, in: path, required: true, schema: { $ref: "#/components/schemas/ProfileIdValue" } }
...
    FileId: { name: fileId, in: path, required: true, schema: { type: string, pattern: '^file_[0-9a-f]{32}$' } }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Required for non-idempotent creates. Scoped by installation, method and canonical path; retained for at least 24 hours. Reuse with a different request fingerprint returns 409.
      schema: { type: string, minLength: 8, maxLength: 128 }
  responses:
    BadRequest:
      description: Invalid request.
      x-error-codes: [validation_failed, invalid_if_match, invalid_cursor]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: Missing or invalid desktop token.
      x-error-codes: [unauthorized]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Resource not found.
      x-error-codes: [resource_not_found, profile_not_found, approval_not_found, clarification_not_found, hermes_state_not_found]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Conflict:
      description: State, revision, idempotency, replay-window or busy conflict.
      x-error-codes: [revision_conflict, idempotency_conflict, idempotency_resource_gone, mcp_server_exists, mcp_config_invalid, capacity_exceeded, unsafe_file_path, memory_storage_drift, profile_delete_conflict, session_busy, session_archived, event_history_expired, hermes_import_source_changed, hermes_import_conflict]
      headers:
        ETag: { $ref: "#/components/headers/ETag" }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    ServiceUnavailable:
      description: A required local service or protected store is unavailable.
      x-error-codes: [secret_storage_unavailable, file_storage_unavailable, skill_storage_unavailable, skill_management_unavailable, operation_storage_unavailable]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    InsufficientStorage:
      description: A bounded local retained-object quota is exhausted; callers must delete no-longer-needed objects before retrying.
      x-error-codes: [file_quota_exceeded]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    SessionStorageUnavailable:
      description: The local session database is unavailable or remained locked past its bounded busy timeout.
      x-error-codes: [session_storage_unavailable, session_storage_busy, hermes_state_unavailable]
      headers:
        Retry-After:
          description: Optional retry delay in seconds for a transient busy database.
          schema: { type: integer, minimum: 0 }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unprocessable:
      description: A required capability/provider is unavailable, or a semantically valid request violates a protected resource constraint.
      x-error-codes: [engine_capability_missing, memory_provider_unsupported, memory_content_blocked, memory_capacity_exceeded, session_search_unavailable, hermes_schema_unsupported, hermes_import_source_invalid, hermes_attachments_require_policy]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    PayloadTooLarge:
      description: Upload or body exceeds the advertised limit.
      x-error-codes: [payload_too_large, hermes_import_too_large]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    UnsupportedMediaType:
      description: The request does not carry the single required media type.
      x-error-codes: [unsupported_media_type, unsupported_file_type]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    ProfileIdValue:
      type: string
      pattern: "^(default|[a-z0-9_][a-z0-9_-]{0,63})$"
...
    HermesImportWarning:
      type: object
      additionalProperties: false
      required: [code, count]
      properties:
        code: { type: string, minLength: 1, maxLength: 128 }
        count: { type: integer, minimum: 1 }
    HermesImportPreview:
      type: object
      additionalProperties: false
      required: [state, adapterId, referenceCommit, schemaVersion, snapshotFingerprint, sessionCount, messageCount, modelUsageRowCount, attachmentCount, rewoundMessageCount, warnings, warningsDropped]
      properties:
        state: { type: string, enum: [absent, ready] }
        adapterId: { type: string, const: hermes-agent-state-v21 }
        referenceCommit: { type: string, minLength: 40, maxLength: 64 }
        schemaVersion: { type: [integer, "null"], minimum: 1 }
        snapshotFingerprint: { type: [string, "null"], pattern: "^[0-9a-f]{64}$" }
        sessionCount: { type: [integer, "null"], minimum: 0 }
        messageCount: { type: [integer, "null"], minimum: 0 }
        modelUsageRowCount: { type: [integer, "null"], minimum: 0 }
        attachmentCount: { type: [integer, "null"], minimum: 0 }
        rewoundMessageCount: { type: [integer, "null"], minimum: 0 }
        warnings: { type: array, maxItems: 256, items: { $ref: "#/components/schemas/HermesImportWarning" } }
        warningsDropped: { type: integer, minimum: 0 }
    HermesImportRequest:
      type: object
      additionalProperties: false
      required: [expectedSnapshotFingerprint, allowAttachmentOmission]
      properties:
        expectedSnapshotFingerprint: { type: string, pattern: "^[0-9a-f]{64}$" }
        allowAttachmentOmission: { type: boolean }
    HermesImportResult:
      type: object
      additionalProperties: false
      required: [importId, profileId, disposition, adapterId, referenceCommit, sourceSchemaVersion, snapshotFingerprint, importedSessionCount, importedMessageCount, importedModelUsageRowCount, omittedAttachmentCount, warnings, warningsDropped]
      properties:
        importId: { type: string }
        profileId: { $ref: "#/components/schemas/ProfileIdValue" }
        disposition: { type: string, enum: [imported, unchanged, replayed] }
        adapterId: { type: string, const: hermes-agent-state-v21 }
        referenceCommit: { type: string, minLength: 40, maxLength: 64 }
        sourceSchemaVersion: { type: integer, const: 21 }
        snapshotFingerprint: { type: string, pattern: "^[0-9a-f]{64}$" }
        importedSessionCount: { type: integer, minimum: 0 }
        importedMessageCount: { type: integer, minimum: 0 }
        importedModelUsageRowCount: { type: integer, minimum: 0 }
        omittedAttachmentCount: { type: integer, minimum: 0 }
        warnings: { type: array, maxItems: 256, items: { $ref: "#/components/schemas/HermesImportWarning" } }
        warningsDropped: { type: integer, minimum: 0 }
    HermesImportConflict:
      type: object
      additionalProperties: false
      required: [code, sourceKeyDigest, targetSessionId]
      properties:
        code: { type: string, enum: [sourceRemoved, sourceChanged, sourceExtended, targetDeleted, targetModified] }
        sourceKeyDigest: { type: string, pattern: "^[0-9a-f]{64}$" }
        targetSessionId: { type: [string, "null"] }
    HermesImportConflictProblem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, code, requestId, retryable, conflictCount, conflicts, conflictsDropped]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, const: 409 }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, const: hermes_import_conflict }
        requestId: { type: string }
        retryable: { type: boolean, const: false }
        conflictCount: { type: integer, minimum: 1 }
        conflicts: { type: array, maxItems: 64, items: { $ref: "#/components/schemas/HermesImportConflict" } }
        conflictsDropped: { type: integer, minimum: 0 }
    FileRef:
      type: object
      additionalProperties: false
      required: [id, name, mimeType, sizeBytes, createdAt]
      properties:
        id: { type: string, pattern: '^file_[0-9a-f]{32}$' }
        name: { type: string, minLength: 1, maxLength: 255 }
        mimeType: { type: string, enum: [application/json, application/octet-stream, application/pdf, application/vnd.openxmlformats-officedocument.spreadsheetml.sheet, application/vnd.openxmlformats-officedocument.wordprocessingml.document, application/x-zip-compressed, application/yaml, application/zip, image/gif, image/jpeg, image/png, image/webp, text/csv, text/markdown, text/plain, text/tab-separated-values, text/yaml] }
        sizeBytes: { type: integer, minimum: 0, maximum: 8388608 }
        createdAt: { type: string, format: date-time }
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, requestId, retryable]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: [string, "null"] }
        instance: { type: [string, "null"], format: uri-reference }
        code:
          type: string
          description: Stable machine-readable code. Each response component's x-error-codes lists known values without closing the set to future additions.
        requestId: { type: string }
        retryable: { type: boolean }